requirements.txt
aiogram>=3.7.0
aiosqlite>=0.20.0
APScheduler>=3.10.4
tzdata>=2024.1
python-dotenv>=1.0.1
flask